+++ tests/src/test_amdpmfreqod.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright 2019 Juan Palacios <jpalaciosdev@gmail.com>
 
-#include "trompeloeil.hpp"
 #include <catch2/catch.hpp>
+#include <catch2/trompeloeil.hpp>
 
 #include "common/commandqueuestub.h"
 #include "common/uintdatasourcestub.h"