+++ tests/src/test_amdpmfreqrange.cpp
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 // Copyright 2021 Juan Palacios <jpalaciosdev@gmail.com>
 
-#include "trompeloeil.hpp"
 #include <catch2/catch.hpp>
+#include <catch2/trompeloeil.hpp>
 
 #include "common/commandqueuestub.h"
 #include "common/vectorstringdatasourcestub.h"